refactor(Leaderboard): relocate dummy data
