+++ packages/client/src/config/constants.tsx
@@ -3,11 +3,3 @@ export enum Paths {
   SignIn = '/sign-in',
   SignUp = '/sign-up',
 }
-
-export const DUMMY_LEADERBOARD = [
-  { place: 1, username: 'Ivan1990', record: 2400, time: 100, matches: 19 },
-  { place: 2, username: 'Marussia', record: 1900, time: 81, matches: 14 },
-  { place: 3, username: 'Me9aTron', record: 500, time: 95, matches: 21 },
-  { place: 4, username: 'SomeBoy', record: 2400, time: 81, matches: 19 },
-  { place: 5, username: 'PVH', record: 1900, time: 95, matches: 14 },
-];
