+++ packages/client/src/pages/Leaderboard/DummyData.tsx
@@ -1,0 +1,7 @@
+export const DUMMY_LEADERBOARD = [
+  { place: 1, username: 'Ivan1990', record: 2400, time: 100, matches: 19 },
+  { place: 2, username: 'Marussia', record: 1900, time: 81, matches: 14 },
+  { place: 3, username: 'Me9aTron', record: 500, time: 95, matches: 21 },
+  { place: 4, username: 'SomeBoy', record: 2400, time: 81, matches: 19 },
+  { place: 5, username: 'PVH', record: 1900, time: 95, matches: 14 },
+];
